refactor(ecs): 重构ECS核心，实现高性能Sparse Set组件系统 (#8)
* feat(ECS): 添加 Component 扩展模块

新增内容:
- ComponentSet64/ComponentSet256 (64/256-bit 组件集合位图)
- ComponentBlockIterator (组件块迭代器)
- ComponentFilterState (组件筛选状态)
- ComponentStorageExtensions (组件存储扩展)
- ComponentTypeId (组件类型ID管理)
- CullingZone (裁剪区域)
- Entity/EntityFlags/EntityRegistry (实体管理完整实现)
- HashCodeUtils (哈希码工具)
- IComponent (组件接口)
- Serialization 模块 (FrameSerializer, HashCodeUtils, IFrameSerializable)

测试:
- ComponentSetBenchmarkTests (性能基准测试)
- ComponentSetCrossPlatformTests (跨平台测试)

修复: 只添加新增文件,不覆盖 main 分支已有 API

* fix(ECS): 移除与 main 分支重复的类型定义

删除重复文件:
- ComponentTypeId.cs (ComponentTypeRegistry 已在 main Frame.cs)
- Entity.cs (main Core/Entity.cs 已存在)
- EntityRegistry.cs (main Core/EntityRegistry.cs 已存在)
- EntityFlags.cs (main EntityMeta.cs 已存在)

保留新增文件:
- ComponentSet64/ComponentSet256/ComponentBlockIterator
- ComponentFilterState/ComponentStorageExtensions
- CullingZone/HashCodeUtils/IComponent
- Serialization 模块
- 测试文件

* fix(ECS): 添加缺失的 using Lattice.Core 引用

ComponentBlockIterator 和 ComponentStorageExtensions 使用了 Entity 类型，
但缺少对 Lattice.Core 命名空间的引用，导致编译失败。

修复内容:
- 添加 using Lattice.Core; 到 ComponentBlockIterator.cs
- 添加 using Lattice.Core; 到 ComponentStorageExtensions.cs

* fix(ECS): 移除与 main 分支 API 不兼容的文件

删除的文件:
- ComponentSet64/ComponentSet256/ComponentFilterState
- ComponentBlockIterator/ComponentStorageExtensions
- ComponentTypeId (重复定义)
- Entity/EntityRegistry/EntityFlags (重复定义)

保留的文件:
- CullingZone (裁剪区域)
- HashCodeUtils (哈希工具)
- IComponent (组件接口)
- Serialization 模块 (FrameSerializer/HashCodeUtils/IFrameSerializable)
- 测试文件 (ComponentSetBenchmarkTests/ComponentSetCrossPlatformTests)

原因: ComponentSet64/ComponentSet256/ComponentFilterState 依赖的 ComponentSet API
与 main 分支的实现不兼容，需要后续重新设计

* fix(ECS): 删除与 main 分支 API 不兼容的测试文件

ComponentSetBenchmarkTests.cs 和 ComponentSetCrossPlatformTests.cs
使用了 feature/ecs-core 分支的 ComponentSet API，
与 main 分支的实现不兼容。

删除的测试文件将在后续重新编写以适应 main 分支 API。

* feat(ECS): FrameSync风格集中式架构，Entity→EntityRef重命名

* feat(ECS): .NET 8 确定性优化 - SIMD/InlineArray/确定性集合

* feat(ecs): 添加确定性测试和性能基准测试套件

- 添加 xUnit 测试项目 (Lattice.Tests.csproj)
- 实现22个确定性验证测试 (EntityRef, ComponentSet, FNV-1a, MemoryLayout)
- 添加8个性能测试 (EntityRef, ComponentSet, Hash, IdMap)
- 修复 ComponentSetNet8 结构大小 (StructLayout Size=64)
- 添加 ComponentSet SIMD 基准测试对比 (Vector512/256 vs Scalar)
- 所有30个测试通过

* refactor(tests): 统一测试目录结构到 scripts/lattice/Tests

- 将原有 godot/Lattice.Tests/ 迁移到 godot/scripts/lattice/Tests/
- 合并原有694个FP测试与新添加的ECS测试
- 新增 ECS 核心测试:
  - EntityRefTests (11个) - 代际索引、内存布局、哈希
  - ComponentSetTests (26个) - 位操作、集合运算、内存布局
  - DeterministicHashTests (10个) - FNV-1a哈希、组合哈希
  - DeterministicIdMapTests (13个) - 确定性字典操作
  - SerializationTests (5个) - 内存序列化验证
  - ECSPerformanceTests (8个) - 性能基准测试
- 更新项目文件启用不安全代码支持
- 移除旧的分散测试目录

总计805个测试全部通过

* fix(ecs): 修复 Component 关键缺陷并优化性能

关键修复：
1. 修复 RemoveComponentInternal - 现在正确调用 storage.Remove() 删除组件数据
2. 禁用懒加载注册 - ComponentTypeId 静态构造函数现在抛出异常，强制显式注册
3. 修复 Query 系统 - Component2/Component3 现在返回 ref 避免值拷贝
4. 修复 FrameBase GCHandle - 移除错误的 Pinned 使用，直接传递非托管缓冲区指针

稳定性改进：
- 添加 IComponentStorage 接口统一存储操作
- 添加结构版本验证 - ComponentEnumerator 检测遍历期间的结构变更
- 添加 BitStream 缓冲区安全检查 - CanRead/CanWrite/VerifyCanRead 方法
- 添加详细的错误消息提升调试体验

性能优化：
- ComponentBlockIterator 移除 FrameBase* 依赖，减少指针解引用
- 优化 EnsureCapacity 缓冲区扩容策略

测试：805个测试全部通过

* refactor(ecs): 重构ECS核心，对标FrameSync实现高性能组件系统

重大重构：
1. 移除双轨制 - 删除Frame/FrameBase双实现，统一为单高性能Frame类
2. Sparse Set存储 - 新Storage<T>使用经典Sparse Set实现（单层稀疏映射）
3. 优化访问路径 - O(1)组件访问，仅需2次内存访问（稀疏数组→密集数组）
4. FrameSync风格Filter - 查询系统自动选择组件数最少的类型作为主遍历源
5. 简化ComponentMask - 512位位图，支持任意组件组合查询
6. 内存优化 - 减少Block管理层级，与FrameSync内存开销对齐

文件变更：
- 新增：Storage.cs, Filter.cs, ComponentFilter.cs
- 重写：Frame.cs, ComponentTypeId.cs, DeterministicHash.cs
- 删除：FrameBase.cs, ComponentStorage.cs, Query.cs等旧实现

测试：694个测试通过

* refactor(ecs): 重构ECS核心，实现高性能Sparse Set组件系统

重大重构：
1. 移除双轨制 - 删除Frame/FrameBase双实现，统一为单高性能Frame类
2. Sparse Set存储 - 新Storage<T>使用经典Sparse Set实现（单层稀疏映射）
3. 优化访问路径 - O(1)组件访问，仅需2次内存访问（稀疏数组→密集数组）
4. 智能Filter - 查询系统自动选择组件数最少的类型作为主遍历源
5. 简化ComponentMask - 512位位图，支持任意组件组合查询
6. 内存优化 - 减少Block管理层级，降低内存开销

文件变更：
- 新增：Storage.cs, Filter.cs, ComponentFilter.cs
- 重写：Frame.cs, ComponentTypeId.cs, DeterministicHash.cs
- 删除：FrameBase.cs, ComponentStorage.cs, Query.cs等旧实现

测试：694个测试通过

* ci: 修复GitHub Actions中的测试项目路径

测试项目已从 godot/Lattice.Tests 移动到 godot/scripts/lattice/Tests
更新所有CI工作流中的路径引用：

- Restore dependencies 步骤路径
- Build 步骤路径
- Test 步骤路径
- Artifact 上传/下载路径
- Cache hashFiles 路径
- trigger paths 路径

---------

+++ godot/scripts/lattice/ECS/Session/Session.cs
@@ -259,7 +259,7 @@ namespace Lattice.ECS.Session
                 return;
             }
 
-            long actualChecksum = frame.CalculateChecksum();
+            long actualChecksum = (long)frame.CalculateChecksum();
             bool success = actualChecksum == expectedChecksum;
 
             if (success)
